Move worker interface to queue pkg
**What**
- Move all common interfaces to a single package to make them easier to
  discover and to reduce the chance of circular imports
