+++ pkg/queue/worker.go
@@ -1,4 +1,4 @@
-package workers
+package queue
 
 import "context"
